+++ common/api/src/main/java/com/dfsek/terra/api/util/provider/DistributorProvider.java
@@ -1,0 +1,4 @@
+package com.dfsek.terra.api.util.provider;
+
+public interface DistributorProvider {
+}
