+++ common/api/src/main/java/com/dfsek/terra/api/util/provider/NoiseProvider.java
@@ -1,6 +1,7 @@
-package com.dfsek.terra.api.util.seeded;
+package com.dfsek.terra.api.util.provider;
 
 import com.dfsek.tectonic.loading.object.ObjectTemplate;
+import com.dfsek.terra.api.util.seeded.NoiseSeeded;
 
 import java.util.function.Supplier;
